feat: update runners, separate configs, add Symbiotic contracts, not working
