+++ blueprints/incredible-squaring-symbiotic/Cargo.toml
@@ -20,6 +20,21 @@ tokio = { workspace = true, default-features = false, features = ["full"] }
 tokio-util = { workspace = true }
 sp-core = { workspace = true }
 subxt-signer = { workspace = true, features = ["sr25519", "subxt", "std"] }
+alloy-contract = { workspace = true }
+alloy-consensus = { workspace = true }
+alloy-json-abi = { workspace = true, features = ["serde_json"] }
+alloy-network = { workspace = true }
+alloy-primitives = { workspace = true }
+alloy-provider = { workspace = true }
+alloy-pubsub = { workspace = true }
+alloy-rpc-types = { workspace = true }
+alloy-rpc-types-eth = { workspace = true }
+alloy-rpc-client = { workspace = true }
+alloy-signer = { workspace = true }
+alloy-signer-local = { workspace = true }
+alloy-sol-types = { workspace = true, features = ["json"] }
+alloy-transport = { workspace = true }
+alloy-transport-http = { workspace = true }
 ark-bn254 = { workspace = true }
 ark-ff = { workspace = true }
 ark-ec = { workspace = true }
@@ -28,7 +43,9 @@ libp2p = { workspace = true }
 ed25519-zebra = { workspace = true, features = ["pkcs8", "default", "der", "std", "serde", "pem"] }
 hex = { workspace = true }
 k256 = { workspace = true }
+serde = { workspace = true }
 serde_json = { workspace = true }
+lazy_static = { workspace = true }
 
 [build-dependencies]
 blueprint-metadata = { workspace = true }
