+++ blueprints/incredible-squaring/src/main.rs
@@ -1,4 +1,5 @@
 use color_eyre::{eyre::eyre, Result};
+use gadget_sdk::config::protocol::TangleInstanceSettings;
 use gadget_sdk::info;
 use gadget_sdk::runners::tangle::TangleConfig;
 use gadget_sdk::runners::BlueprintRunner;
@@ -12,17 +13,17 @@ async fn main() {
 
     info!("Starting the event watcher for {} ...", signer.account_id());
 
+    let tangle_settings = env.protocol_specific.tangle()?;
+    let TangleInstanceSettings { service_id, .. } = tangle_settings;
     let x_square = blueprint::XsquareEventHandler {
-        service_id: env.service_id.unwrap(),
+        service_id: *service_id,
         context: blueprint::MyContext,
         client,
         signer,
     };
 
     info!("~~~ Executing the incredible squaring blueprint ~~~");
-    let tangle_config = TangleConfig {
-        price_targets: Default::default(),
-    };
+    let tangle_config = TangleConfig::default();
     BlueprintRunner::new(tangle_config, env)
         .add_job(x_square)
         .run()
